Jojos: Revamp Jojos layout
Rewrite how we pull in Jojos data.  The new layout also simplifies updates and maintenance to Jojos.

Minor updates around the codebase to stay in sync.

+++ palmod/Game/GameLoad.h
@@ -10,7 +10,7 @@ private:
     UINT16(*GetRuleCtr)(void);
     void (*ResetRuleCtr)(void);
     sFileRule(*GetNextRule)(void);
-    sFileRule(*GetRule)(int);
+    sFileRule(*GetRule)(UINT16);
     CGameClass* CreateGame(int nGameFlag, int nExtraGameData = 0);
 
     CString szLoadSaveStr = "";
