+++ palmod/Game/Game_MVC_A.h
@@ -15,7 +15,7 @@ private:
     int nCurrPalOffs = 0;
     int nCurrPalSz = 0;
 
-    void GetPalOffsSz(int nUnitId, int nPalId);
+    void GetPalOffsSz(UINT16 nUnitId, UINT16 nPalId);
 
     void InitDataBuffer();
     void ClearDataBuffer();
@@ -30,27 +30,27 @@ public:
     static CDescTree MainDescTree;
 
     static CDescTree InitDescTree();
-    static sFileRule GetRule(int nUnitId);
+    static sFileRule GetRule(UINT16 nUnitId);
 
     //Extra palette function
-    static int GetExtraCt(int nUnitId, BOOL bCountVisibleOnly = FALSE);
-    static int GetExtraLoc(int nUnitId);
-    static int GetBasicAmt(int nUnitId);
+    static int GetExtraCt(UINT16 nUnitId, BOOL bCountVisibleOnly = FALSE);
+    static int GetExtraLoc(UINT16 nUnitId);
+    static int GetBasicAmt(UINT16 nUnitId);
 
     static void LoadExtraFile();
 
     //Normal functions
     CDescTree* GetMainTree();
 
-    static int GetPalCt(int nUnitId);
+    static int GetPalCt(UINT16 nUnitId);
 
-    void CreateDefPal(sDescNode* srcNode, int nSepId);
-    BOOL LoadFile(CFile* LoadedFile, int nUnitId = 0);
-    BOOL SaveFile(CFile* SaveFile, int nUnitId = 0);
+    void CreateDefPal(sDescNode* srcNode, UINT16 nSepId);
+    BOOL LoadFile(CFile* LoadedFile, UINT16 nUnitId = 0);
+    BOOL SaveFile(CFile* SaveFile, UINT16 nUnitId = 0);
     BOOL UpdatePalImg(int Node01 = -1, int Node02 = -1, int Node03 = -1, int Node04 = -1);
 
-    COLORREF* CreatePal(int nUnitId, int nPalId);
-    BOOL CreateExtraPal(int nUnitId, int nPalId);
+    COLORREF* CreatePal(UINT16 nUnitId, UINT16 nPalId);
+    BOOL CreateExtraPal(UINT16 nUnitId, UINT16 nPalId);
 
     void UpdatePalData();
