+++ palmod/Game/Game_SFIII3_D.h
@@ -15,7 +15,7 @@ private:
 
     UINT16*** pppDataBuffer = nullptr;
 
-    void GetPalOffsSz(int nUnitId, int nPalId);
+    void GetPalOffsSz(UINT16 nUnitId, UINT16 nPalId);
 
     void InitDataBuffer();
     void ClearDataBuffer();
@@ -33,27 +33,27 @@ public:
 
     static UINT16 uRuleCtr;
     static sFileRule GetNextRule();
-    static sFileRule GetRule(int nUnitId);
+    static sFileRule GetRule(UINT16 nUnitId);
 
     static UINT16 GetRuleCtr() { return uRuleCtr; };
     static void ResetRuleCtr() { uRuleCtr = 0; };
 
     //Extra palette function
-    static int GetBasicAmt(int nUnitId);
+    static int GetBasicAmt(UINT16 nUnitId);
 
     static void LoadExtraFile();
 
     //Normal functions
     CDescTree* GetMainTree();
 
-    static int GetPalCt(int nUnitId);
+    static int GetPalCt(UINT16 nUnitId);
 
-    void CreateDefPal(sDescNode* srcNode, int nSepId);
-    BOOL LoadFile(CFile* LoadedFile, int nUnitId = 0);
-    BOOL SaveFile(CFile* SaveFile, int nUnitId = 0);
+    void CreateDefPal(sDescNode* srcNode, UINT16 nSepId);
+    BOOL LoadFile(CFile* LoadedFile, UINT16 nUnitId = 0);
+    BOOL SaveFile(CFile* SaveFile, UINT16 nUnitId = 0);
     BOOL UpdatePalImg(int Node01 = -1, int Node02 = -1, int Node03 = -1, int Node04 = -1);
 
-    COLORREF* CreatePal(int nUnitId, int nPalId);
+    COLORREF* CreatePal(UINT16 nUnitId, UINT16 nPalId);
 
     void UpdatePalData();
