+++ palmod/Game/Game_XMVSF_A.h
@@ -11,7 +11,7 @@ private:
 
     UINT16*** pppDataBuffer = nullptr;
 
-    void GetPalOffsSz(int nUnitId, int nPalId);
+    void GetPalOffsSz(UINT16 nUnitId, UINT16 nPalId);
 
     void InitDataBuffer();
     void ClearDataBuffer();
@@ -24,28 +24,28 @@ public:
     static CDescTree MainDescTree;
 
     static CDescTree InitDescTree();
-    static sFileRule GetRule(int nUnitId);
+    static sFileRule GetRule(UINT16 nUnitId);
 
     //Extra palette function
-    static int GetBasicAmt(int nUnitId);
-    static int GetPalCt(int nUnitId);
+    static int GetBasicAmt(UINT16 nUnitId);
+    static int GetPalCt(UINT16 nUnitId);
 
     static void LoadExtraFile();
 
     //Normal functions
     CDescTree* GetMainTree();
 
-    void CreateDefPal(sDescNode* srcNode, int nSepId);
-    BOOL LoadFile(CFile* LoadedFile, int nUnitId = 0);
-    BOOL SaveFile(CFile* SaveFile, int nUnitId = 0);
+    void CreateDefPal(sDescNode* srcNode, UINT16 nSepId);
+    BOOL LoadFile(CFile* LoadedFile, UINT16 nUnitId = 0);
+    BOOL SaveFile(CFile* SaveFile, UINT16 nUnitId = 0);
     BOOL UpdatePalImg(int Node01 = -1, int Node02 = -1, int Node03 = -1, int Node04 = -1);
 
-    COLORREF* CreatePal(int nUnitId, int nPalId);
+    COLORREF* CreatePal(UINT16 nUnitId, UINT16 nPalId);
 
     void UpdatePalData();
 
     void FlushUnitFile() { safe_delete(rgFileChanged); };
     void PrepUnitFile() { if (!rgFileChanged) { rgFileChanged = new UINT16; } };
 
-    int GetLocalAmt(int nUnitId) { return (XMVSF_A_UNITLOC[nUnitId + 1] - XMVSF_A_UNITLOC[nUnitId]) / 0x20; };
+    int GetLocalAmt(UINT16 nUnitId) { return (XMVSF_A_UNITLOC[nUnitId + 1] - XMVSF_A_UNITLOC[nUnitId]) / 0x20; };
 };
